day03/pyproject.toml
[project]
name = "day03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-levenshtein>=0.27.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
]
